requirements.txt
aiofiles==0.7.0
anyio==3.3.4
asgiref==3.4.1
click==8.0.3
fastapi==0.70.0
flake8==4.0.1
h11==0.12.0
idna==3.3
mccabe==0.6.1
numpy==1.21.3
opencv-python==4.5.4.58
Pillow==8.4.0
pycodestyle==2.8.0
pydantic==1.8.2
pyflakes==2.4.0
pytesseract==0.3.8
python-decouple==3.5
python-multipart==0.0.5
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
typing-extensions==3.10.0.2
uuid==1.30
uvicorn==0.15.0
